Prototype Cura bounding box post-processor for dynamic ABL
This is a prototype post-processing plugin for the Cura slicer.  To
use, copy it to your Cura plugin directory (e.g. ~/.cura/plugin) and
add it the list of plugins for an object.

This is a prototype for finding the bounding box on the first layer and
printing it out.  It isn’t a fully functional plugin, yet!
See this page for turning on debugging in Cura so that you can see the
output:
http://wiki.ultimaker.com/How_to_write_a_Cura_plugin#Debugging_your_plug
in